requirements.txt
django
djangorestframework
django-cors-middleware
requests
boto3
django-storages
